% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooksd_OPT.R
\name{cooksd_OPT}
\alias{cooksd_OPT}
\title{Function to calculate cooks distance}
\usage{
cooksd_OPT(model)
}
\description{
This function calculates cooks distance by dividing the residuals squared by number of parameters(p) multiplied by our MSE; this is multiplied by our leverage values divided by (1 - our leverage values)squared
cooks distance = (residuals^2/(p*MSE))*levrage values/(1-leverage values)^2
}
\examples{
cooksd_OPT(modelx)
}
