% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid_OPT.R
\name{resid_OPT}
\alias{resid_OPT}
\title{Function to compute basic residuals of linear regression}
\usage{
resid_OPT(model)
}
\description{
This function calculates the residuals of a specified linear model by subtracting the fitted value from y
Residual = y - fitted_values
}
\examples{
resid_OPT(modelx)
}
