% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rstand_OPT.R
\name{Rstand_OPT}
\alias{Rstand_OPT}
\title{Function to calculate standardized residuals}
\usage{
Rstand_OPT(model)
}
\description{
This function calculates the standardized residuals by dividing the residuals by the residual standard error multiplied by the square rooot of 1 - our levarage values
Standard Residuals = Residuals/(estimated residual standard error*sqrt(1-leverage values)
}
\examples{
Rstand_OPT(modelx)
}
