% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vif_OPT.R
\name{vif_OPT}
\alias{vif_OPT}
\title{Function to calculate VIF (Variance Inflation Factor)}
\usage{
vif_OPT(model)
}
\description{
This function calculates the vif by dividing 1 by 1- R-squared value for ith predictor
cooks distance = 1/(1-R^2)
}
\examples{
vif_opt(modelx)
}
